+++ arborio/cableio_error.cpp
@@ -1,0 +1,13 @@
+#include <iostream>
+
+#include <arbor/s_expr.hpp>
+#include <arborio/cableio_error.hpp>
+
+namespace arborio {
+// Errors
+cableio_parse_error::cableio_parse_error(const std::string& msg, const arb::src_location& loc):
+    arb::arbor_exception(msg+" at :"+std::to_string(loc.line)+":"+std::to_string(loc.column))
+{}
+cableio_unexpected_symbol::cableio_unexpected_symbol(const std::string& sym, const arb::src_location& loc):
+    cableio_parse_error("Unexpected symbol "+sym, loc) {}
+} // namespace arborio
